:truck: Fix directory structure
Resolves: #3
