+++ src/main/com/copilot/yagl4j/GamePadAxis.java
