+++ src/main/com/copilot/yagl4j/GamePadManager.java
