+++ src/native/CMakeLists.txt
@@ -12,8 +12,8 @@ PKG_SEARCH_MODULE(SDL2IMAGE REQUIRED SDL2_image>=2.0.0)
 set(SOURCE_FILES library.cpp library.h)
 set(CMAKE_CXX_FLAGS ${SDL_CFLAGS} -pthread)
 
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/build)
-set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/build)
+set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/../build)
+set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_SOURCE_DIR}/../build)
 
 include_directories(${SDL2_INCLUDE_DIRS} ${SDL2IMAGE_INCLUDE_DIRS} ${JNI_INCLUDE_DIRS})
 link_libraries(${SDL2_LIBRARIES} ${SDL2IMAGE_LIBRARIES} ${JNI_LIBRARIES})
